client: Add error log for missing health package import during health check (#8595)
Addresses: https://github.com/grpc/grpc-go/issues/8590

When using the health producer for health checks, and the health package
is not imported by the application, a no op health producer is used
without logging any errors. This PR adds an error log similar to the one
for the old health checks started by the subchannel.


https://github.com/grpc/grpc-go/blob/e35080456c7071b6d55887080ef78809c163640c/clientconn.go#L1475-L1481

RELEASE NOTES: N/A