xdsclient: fix the flaky ADS stream restart test (#8631)
The ADS stream restart test can be flaky for the following reason:
- It requests a CDS resource and unrequests it before the stream breaks.
- And then once the stream restarts, it verifies that this resource is
not requested again.
- But the ACK for this resource may or may not be received at the
management server before the stream breaks. This can falsely cause the
test to conclude that the request was re-requested after the restart.

This PR changes the test in the following ways:
- Use a single resource
- Verify ACK before the stream is restarted

Ran a million times without flakes on Forge.

RELEASE NOTES: NONE