client: move tests that rely on `Dial` to a separate file (#8707)
This PR only moves some tests around, does not make any changes to the
tests.

Moving tests that rely on the behavior of `Dial` to a separate file
makes them easier to find and work with.

I will be cleaning up these tests a little in a follow up PR, and will
be adding a new test for https://github.com/grpc/grpc-go/issues/7686

RELEASE NOTES: none