Cherrypick #8702 to v1.77.x (#8709)
https://github.com/grpc/grpc-go/pull/8548 changed the grpc status code
returned on missing grpc-status header to Internal. This PR changes it
back to Unknown as that is the expected behavior.

See
[grpc/grpc@master/doc/statuscodes.md](https://github.com/grpc/grpc/blob/master/doc/statuscodes.md)
for more details. Here Unknown is defined as follows:
```
Unknown error. For example, this error may be returned when a Status value received 
from another address space belongs to an error space that is not known in this 
address space. Also errors raised by APIs that do not return enough error information
may be converted to this error.
```

RELEASE NOTES:
- transport/client : Return status code Unknown on missing grpc-status.

Co-authored-by: Madhav Bissa <48023579+mbissa@users.noreply.github.com>