xdsclient/xdsresource: add AutoHostRewrite and Endpoint Hostname support (#8728)
This PR implements the validation logic and extracting per endpoint
Hostname attributes from xDS resources for [gRFC
A81](https://github.com/grpc/proposal/blob/master/A81-xds-authority-rewriting.md)

### Key Changes:

1.  **RDS Resource Validation :**
* The boolean value of `RouteAction.auto_host_rewrite` is extracted from
the RDS resource and stored in route struct
* This field is only set to `true` in the parsed route struct if the
`trusted_xds_server` option is present in the `ServerConfig` and the
global environment variable for authority overriding is enabled.

2.  **EDS Resource Validation:**
* The `Endpoint.hostname` field is extracted from the EDS resource and
will be stored as a `hostname` string in parsed endpoint struct. It will
be changed to be an per-endpoint resolver attribute in a follow-up PR.

RELEASE NOTES: None