client: Change connectivity state to CONNECTING when creating the name resolver (#8710)
Fixes https://github.com/grpc/grpc-go/issues/7686

#### Current Behavior
- When client exits IDLE and creates the name resolver, it stays in IDLE
until the connectivity state is set by the LB policy.
- When exiting IDLE mode (because of `Connect` being called or because
of an RPC), if name resolver creation fails, we stay in IDLE.

#### New Behavior
- When the client exits IDLE and creates the name resolver, it moves to
CONNECTING. Moving forward, the connectivity state will be set by the LB
policy.
- When exiting IDLE mode (because of `Connect` being called or because
of an RPC), we have already moved to CONNECTING (because of the previous
bullet point). If name resolver creation fails, we will move to
TRANSIENT_FAILURE and start the idle timer and move back to IDLE when
the timer fires

#### Implementation details:
- The client channel now treats resolver build errors encountered during
exiting IDLE identically to resolver errors received prior to valid
updates.
- `Build` uses a new unsafe API on the idleness manager to mark the
channel as exited IDLE.
- The idleness Manager invokes the channel's `ExitIdleMode` (which now
does not return an error) and updates internal state to reflect that it
is no longer in IDLE.
- `OnFinish` call options are now invoked even if stream creation fails
during an RPC. This fulfills the guarantee for these options and ensures
the idleness Manager’s `activeCallsCount` remains accurate.

RELEASE NOTES:
- client: Change connectivity state to CONNECTING when creating the name
resolver (as part of exiting IDLE).
- client: Change connectivity state to TRANSIENT_FAILURE if name
resolver creation fails (as part of exiting IDLE).
- client: Change connectivity state to IDLE after idle timeout expires
even when current state is TRANSIENT_FAILURE.
- client: Fix a bug that resulted in `OnFinish` call option not being
invoked for RPCs where stream creation failed.